services/query-service/requirements.txt
fastapi==0.116.0
pydantic==2.11.7
python-dotenv==1.1.1
pymongo==4.13.2
requests==2.32.4
redis==6.2.0
uvicorn==0.35.0
anyio==4.9.0
starlette==0.46.2
typing_extensions==4.14.1
google-genai==1.25.0